PLUGIN/plugin_backend/requirements.txt
playwright
requests
beautifulsoup4
langchain
langchain-community
langchain-openai
python-dotenv
openai
Flask
